+++ CHANGELOG.md
@@ -3,6 +3,7 @@ See also [https://github.com/stkb/vscode-rewrap/releases](https://github.com/stk
 ## Unreleased
 
 - Fix for settings changes not applying until an editor restart (#99)
+- Modify "Rewrap at custom column" command to make "unwrapping" easier (#107)
 - AsciiDoc: Fix only comments being wrapped (#196)
 - R: Allow `#'` prefix for ROxygen comments (full ROxygen support still to come)
   (#181)